# Perf: optimize actual_buffer_size to use only data buffer capacity for coalesce