--- a/arrow-select/src/coalesce/byte_view.rs
+++ b/arrow-select/src/coalesce/byte_view.rs
@@ -286,7 +286,8 @@ impl<B: ByteViewType> InProgressArray for InProgressByteViewArray<B> {
                 let ideal_buffer_size = s.total_buffer_bytes_used();
                 // We don't need to use get_buffer_memory_size here, because we gc is mainly for
                 // data buffers, not views and nulls.
-                let actual_buffer_size = s.data_buffers().iter().map(|b| b.capacity()).sum::<usize>();
+                let actual_buffer_size =
+                    s.data_buffers().iter().map(|b| b.capacity()).sum::<usize>();
                 // copying strings is expensive, so only do it if the array is
                 // sparse (uses at least 2x the memory it needs)
                 let need_gc =