@zhuqi-lucas 2025-07-20
But the test failed, i am not sure if this is a reasonable optimization. 🤔