@zhuqi-lucas 2025-07-23
> Thank you @zhuqi-lucas -- I double checked and indeed since we are compacting the buffers, it makes sense to use the buffer size rather than the total memory size (that also includes the views and null buffers)

Thank you @alamb for review and double checking!