@alamb 2025-07-23
🤖: Benchmark completed

<details><summary>Details</summary>
<p>


```
group                                                          main                                   optimize_actual_buffer_size
-----                                                          ----                                   ---------------------------
concat 1024 arrays boolean 4                                   1.01     28.0±0.13µs        ? ?/sec    1.00     27.8±0.04µs        ? ?/sec
concat 1024 arrays i32 4                                       1.01     14.2±0.09µs        ? ?/sec    1.00     14.1±0.06µs        ? ?/sec
concat 1024 arrays str 4                                       1.01     56.9±0.44µs        ? ?/sec    1.00     56.1±0.56µs        ? ?/sec
concat boolean 1024                                            1.30    451.8±0.45ns        ? ?/sec    1.00    348.2±0.36ns        ? ?/sec
concat boolean 8192 over 100 arrays                            1.00     51.0±0.08µs        ? ?/sec    1.00     50.9±0.07µs        ? ?/sec
concat boolean nulls 1024                                      1.16    786.7±4.89ns        ? ?/sec    1.00    680.5±1.17ns        ? ?/sec
concat boolean nulls 8192 over 100 arrays                      1.00    109.6±0.19µs        ? ?/sec    1.00    109.7±0.17µs        ? ?/sec
concat fixed size lists                                        1.00   709.2±32.96µs        ? ?/sec    1.00   707.8±32.05µs        ? ?/sec
concat i32 1024                                                1.00    437.1±1.21ns        ? ?/sec    1.00    437.0±0.89ns        ? ?/sec
concat i32 8192 over 100 arrays                                1.00    218.1±9.45µs        ? ?/sec    1.01    220.9±5.36µs        ? ?/sec
concat i32 nulls 1024                                          1.13    760.3±2.11ns        ? ?/sec    1.00    675.6±4.53ns        ? ?/sec
concat i32 nulls 8192 over 100 arrays                          1.01    283.1±7.66µs        ? ?/sec    1.00    279.3±9.60µs        ? ?/sec
concat str 1024                                                1.00     13.5±1.09µs        ? ?/sec    1.02     13.8±1.24µs        ? ?/sec
concat str 8192 over 100 arrays                                1.00    107.8±1.04ms        ? ?/sec    1.00    107.8±0.84ms        ? ?/sec
concat str nulls 1024                                          1.00      6.6±0.52µs        ? ?/sec    1.02      6.7±0.84µs        ? ?/sec
concat str nulls 8192 over 100 arrays                          1.01     53.9±0.46ms        ? ?/sec    1.00     53.4±0.37ms        ? ?/sec
concat str_dict 1024                                           1.07      3.0±0.01µs        ? ?/sec    1.00      2.8±0.02µs        ? ?/sec
concat str_dict_sparse 1024                                    1.01      7.0±0.02µs        ? ?/sec    1.00      7.0±0.03µs        ? ?/sec
concat struct with int32 and dicts size=1024 count=2           1.00      6.7±0.15µs        ? ?/sec    1.00      6.7±0.12µs        ? ?/sec
concat utf8_view  max_str_len=128 null_density=0               1.01     77.8±0.34µs        ? ?/sec    1.00     77.3±0.38µs        ? ?/sec
concat utf8_view  max_str_len=128 null_density=0.2             1.00     84.1±0.43µs        ? ?/sec    1.00     84.2±0.39µs        ? ?/sec
concat utf8_view  max_str_len=20 null_density=0                1.01     77.9±0.39µs        ? ?/sec    1.00     77.5±0.38µs        ? ?/sec
concat utf8_view  max_str_len=20 null_density=0.2              1.00     84.2±0.40µs        ? ?/sec    1.00     83.9±0.39µs        ? ?/sec
concat utf8_view all_inline max_str_len=12 null_density=0      1.00     40.9±4.43µs        ? ?/sec    1.13     46.2±3.80µs        ? ?/sec
concat utf8_view all_inline max_str_len=12 null_density=0.2    1.00     46.9±3.28µs        ? ?/sec    1.20     56.0±5.19µs        ? ?/sec
```


</p>
</details>