@alamb 2025-07-23
🤖 `./gh_compare_arrow.sh` [Benchmark Script](https://github.com/alamb/datafusion-benchmarking/blob/main/gh_compare_arrow.sh) Running
Linux aal-dev 6.11.0-1016-gcp #16~24.04.1-Ubuntu SMP Wed May 28 02:40:52 UTC 2025 x86_64 x86_64 x86_64 GNU/Linux
Comparing optimize_actual_buffer_size (2df2acf6e044fcf1575f60f7764fb31f171e15bf) to 82821e574df7e699c7a491da90c54429a5a439e9 [diff](https://github.com/apache/arrow-rs/compare/82821e574df7e699c7a491da90c54429a5a439e9..2df2acf6e044fcf1575f60f7764fb31f171e15bf)
BENCH_NAME=filter_kernels
BENCH_COMMAND=cargo bench --features=arrow,async,test_common,experimental --bench filter_kernels 
BENCH_FILTER=
BENCH_BRANCH_NAME=optimize_actual_buffer_size
Results will be posted here when complete