@zhuqi-lucas 2025-07-23
This is matching my local result, 2x faster for high select!

```rust
filter: single_utf8view, 8192, nulls: 0, selectivity: 0.8                            2.62      3.7±0.02ms        ? ?/sec    1.00   1408.1±4.77µs        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.001                        1.00     90.0±0.27ms        ? ?/sec    1.01     90.8±0.23ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.01                         1.00     11.2±0.03ms        ? ?/sec    1.01     11.3±0.04ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.1                          1.00      5.6±0.34ms        ? ?/sec    1.06      6.0±0.41ms        ? ?/sec
filter: single_utf8view, 8192, nulls: 0.1, selectivity: 0.8                          1.63      6.2±0.02ms        ? ?/sec    1.00      3.8±0.01ms        ? ?/sec
```

> 🤖: Benchmark completed
> 
> Details